pyproject.toml
[project]
name = "ubdc-cctv"
version = "0.2.2"
description = "CCTV backend for the University of Glasgow's Urban Big Data Centre"
authors = [{ name = "nikos ves", email = "nick.ves@glasgow.ac.uk" }]
requires-python = ">=3.10, <4"
dependencies = [
    "django>=4.2.0,<5",
    "djangorestframework>=3.13.1,<4",
    "Markdown>=3.3.6,<4",
    "django-filter>=25,<26",
    "python-dateutil>=2.8.1,<3",
    "pyyaml>=6.0.0,<7",
    "uvicorn>=0.17.0,<1",
    "whitenoise[brotli]<7.0",
    "uritemplate>=4.1.1,<5",
    "django-storages>=1.12.3,<2",
    "django-cors-headers>=4,<5",
    "pydantic>=1.10.5,<2",
    "drf-spectacular>=0.29.0",
]

[dependency-groups]
dev = [
    "freezegun>=1.2.2,<2",
    "pytest>=7.1.1,<8",
    "pytest-django>=4.5.2,<5",
    "pytest-drf>=1.1.3,<2",
]

[tool.pytest.ini_options]
# addopts = "-ra"
DJANGO_SETTINGS_MODULE = "cctv_core.settings"
pythonpath = [ "cctv_api" ]

[tool.ruff]
line-length = 120
extend-exclude = [
    ".venv",
]

[tool.ruff.lint]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"
